Move ray query helpers to be used by pipelines into the shared module.
